Lmaxmem must be computed as long long.

+++ src/hhblits.C
@@ -1236,8 +1236,8 @@ void perform_realign(char *dbfiles[], int ndb)
   int N_aligned=0;
 
   // Longest allowable length of database HMM (backtrace: 5 chars, fwd: 1 double, bwd: 1 double 
-  int Lmaxmem=(int)((par.maxmem-0.5)*1024*1024*1024)/(2*sizeof(double)+8)/q->L/bins;
-  int Lmax=0;      // length of longest HMM to be realigned
+  long long Lmaxmem=((par.maxmem-0.5)*1024*1024*1024)/(2*sizeof(double)+8)/q->L/bins;
+  long long Lmax=0;      // length of longest HMM to be realigned
     
   par.block_shading->Reset();
   while (!par.block_shading->End())
